[#90545416] Add spec for editing car
Added spec for editing a car
Added support files and methods to pass spec
Edit and create has some redundancy in the form; need to refactor
this out to a partial.
Did some refactoring of create specs
Some redundancy in create and edit specs for filling in new car
form; perhaps refactor that out to a helper?